First set of changes from code review.
Specifically:

`lxml` needs to be added to `requirements.txt`
`parser/parser.py:112`: What is `ns` in the name of this function? Being in the name of the function makes me think it is important, but scares me that I can't work out what it is.
`parser/parser.py:376`: This seems to be a misleading name and comment - as far as I can tell, this is the entry to parsing the whole set of documents, not just the region map
`parser/parser.py:34`: This `dita_soup` variable is created here, used briefly to create an image tag, and then assigned to again at `parser/parser.py:50` - and first assignment uses `lxml-xml` and the second
just `xml`. I suspect the second definition can be moved up to replace the first one, but I'm not sure
`parser/parser_utils.py:77`: This actually removes _multiple_ leading slashes
`parser/parser.py:117`: Comment outdated, this just sets up the HTML parsing
`parser/parser_utils.py:18`: This function can be replaced with `os.makedirs` with the argument `exist_ok=True`
`parser/parser_utils.py:36`: This could be simplified by replacing the first if with `if not filenames: filenames = os.listdir(source_dir)` and then just having one lot of copying commands (reduces code
duplication)
`parser/parser_utils.py:73`: This function probably isn't needed, I don't think it is any simpler than the standard python version of `s.replace('x', 'y')`