More changes from Robin's code review.
Specifically:
`parser/parser.py:270`: tr_count in this for loop is unused - if you're not using it then you can remove the enumerate call. Same applies a couple of lines later.
`parser/parser.py:343`: It looks like this call to `get_files_in_path` can be removed and you can just pass an empty list of files to `copy_files`, and then it'll copy all the files in the source folder. The
same applies at `parser/parser.py:210`. That seems to be the only two places that `get_files_in_path` is used, so it can probably be removed after that.
`parser/html_to_dita.py:51`: This if statement seems unnecessary - you can just do the assignment and if the value is the same as what it's being replaced with then it'll stay the same. The only reason not to
do this is if there is a performance impact, which I don't think there would be here.
`parser/html_to_dita.py:49`: This actually replaces it with whatever is in div_replacement, so the comment is misleading
`parser/html_to_dita.py:58`: This could do with a numbered header as it seems to be a different transformation, and not part of number 1
`parser/html_to_dita.py:121`: This comment and the one below it are out of step with the code - they should be h2 and h3 respectively. In fact, this code could be replaced with another loop that loops over
"h1", "h2" and "h3".
`parser/html_to_dita.py:144`: The repeated code here could be put in a for loop too
`parser/parser_utils.py:83`: Not sure why you've got a definition for `__all__` here or in various other utility files - that's normally just in an `__init__.py`, and importing should work automatically
without `__all__` in a normal Python file.