Fixed more issues from Robin's code review
Specifically:
parser/reference_files.py:9`: This function needs a clearer docstring - it says it just converts a HTML file to DITA, but that's a description of the whole project - and more specifically of the `htmlToDITA`
function. The docstring should explain what a non-class file is, and how it parses it.
`parser/html_to_dita.py:262`: This is just checking whether it is one character long, rather than checking if it's actually just a newline. I doubt there are many single character bits of text, but it is
probably worth checking for `\n` here instead/as well.
A useful new function could be something like `write_prettified_xml` which takes a dita_soup variable and does `parser/reference_files.py:76` to `parser/reference_files.py:79` returning the filename - this
could be used in various places and would reduce a bit of duplication.